Fix: sizes

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int l;
 	float m;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(i));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(j));
-	printf("Size of a long int is: %lu byte(s)\n", (unsigned long)sizeof(k));
-	printf("Size of a long long int is: %lu byte(s)\n", (unsigned long)sizeof(l));
-	printf("The size of a float is: %lu byte(s)\n", (unsigned long)sizeof(m));
+	printf("Size of a char: %lu byte(s)\n", sizeof(i));
+	printf("Size of an int: %lu byte(s)\n", sizeof(j));
+	printf("Size of a long int is: %lu byte(s)\n", sizeof(k));
+	printf("Size of a long long int is: %lu byte(s)\n", sizeof(l));
+	printf("The size of a float is: %lu byte(s)\n", sizeof(m));
 	return (0);
 }
